fix: add get-block-command
